Make sure shard_elements is < 2^30, and optimize/cleanup accordingly.
This ensures that expressions like (blockIdx.x * blockDim.x + threadIdx.x) where each thread owns an element are safe.

+++ utilities_cuda.cuh
@@ -355,7 +355,7 @@ struct plane2split {
 	completion plane(double2* planar, cudaStream_t stream = 0) const { return plane(planar, stream, stream); }
 
 	const uint16_t chainlen, shard_copies;
-	const size_t elements;
+	const uint32_t elements;
 	const cudalist<double> real_transposed;
 	double* const img_transposed;
 private:
